requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pymysql==1.1.0
ultralytics==8.1.31
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0
tqdm==4.66.2
pydantic==2.6.4
python-dotenv==1.0.1